Add files via upload
This is part of a graduation project I wrote for a class in python. The program which tests which method wins more often in kalaha. One method is to choose bowls randomly, the other is to choose the bowl with most rocks in it. The user can choose the number of simulations that are being played, and the number of rocks in each bowl from the beginning (3 or 6). The instructions are in Swedish.